talk_to_the_future/requirements.txt
colorama>=0.4.6
questionary>=2.1.0
PyNaCl==1.5.0